Remove experimentReducer - experimentReducer removed from project - Now experimentsReducer includes all actions of experiment, using a selector to fetch a single experiment
