--- src/store/experiment/actionTypes.js
@@ -1,85 +1,0 @@
-// ACTION TYPES
-// REQUEST
-// fetch experiment request
-const FETCH_EXPERIMENT_REQUEST = 'FETCH_EXPERIMENT_REQUEST';
-// edit experiment name request
-const EDIT_EXPERIMENT_NAME_REQUEST = 'EDIT_EXPERIMENT_NAME_REQUEST';
-// deploy experiment request
-const DEPLOY_EXPERIMENT_REQUEST = 'DEPLOY_EXPERIMENT_REQUEST';
-// create experiment request
-const CREATE_EXPERIMENT_REQUEST = 'CREATE_EXPERIMENT_REQUEST';
-// delete experiment request
-const DELETE_EXPERIMENT_REQUEST = 'DELETE_EXPERIMENT_REQUEST';
-// fetch experiment deploy status request
-const FETCH_EXPERIMENT_DEPLOY_STATUS_REQUEST =
-  'FETCH_EXPERIMENT_DEPLOY_STATUS_REQUEST';
-
-// SUCCESS
-// fetch experiment success
-const FETCH_EXPERIMENT_SUCCESS = 'FETCH_EXPERIMENT_SUCCESS';
-// edit experiment name success
-const EDIT_EXPERIMENT_NAME_SUCCESS = 'EDIT_EXPERIMENT_NAME_SUCCESS';
-// deploy experiment success
-const DEPLOY_EXPERIMENT_SUCCESS = 'DEPLOY_EXPERIMENT_SUCCESS';
-// create experiment success
-const CREATE_EXPERIMENT_SUCCESS = 'CREATE_EXPERIMENT_SUCCESS';
-// delete experiment success
-const DELETE_EXPERIMENT_SUCCESS = 'DELETE_EXPERIMENT_SUCCESS';
-// fetch experiment deploy status success
-const FETCH_EXPERIMENT_DEPLOY_STATUS_SUCCESS =
-  'FETCH_EXPERIMENT_DEPLOY_STATUS_SUCCESS';
-
-// FAIL
-// fetch experiment fail
-const FETCH_EXPERIMENT_FAIL = 'FETCH_EXPERIMENT_FAIL';
-// deploy experiment fail
-const DEPLOY_EXPERIMENT_FAIL = 'DEPLOY_EXPERIMENT_FAIL';
-// create experiment fail
-const CREATE_EXPERIMENT_FAIL = 'CREATE_EXPERIMENT_FAIL';
-// delete experiment fail
-const DELETE_EXPERIMENT_FAIL = 'DELETE_EXPERIMENT_FAIL';
-// fetch experiment deploy status fail
-const FETCH_EXPERIMENT_DEPLOY_STATUS_FAIL =
-  'FETCH_EXPERIMENT_DEPLOY_STATUS_FAIL';
-
-// COMMON
-// training experiment succeeded
-const TRAINING_EXPERIMENT_SUCCEEDED = 'TRAINING_EXPERIMENT_SUCCEEDED';
-// training experiment not succeeded
-const TRAINING_EXPERIMENT_NOT_SUCCEEDED = 'TRAINING_EXPERIMENT_NOT_SUCCEEDED';
-
-//DUPLICATE
-const DUPLICATE_EXPERIMENT_FAIL = 'DUPLICATE_EXPERIMENT_FAIL';
-
-// EXPORT
-export default {
-  // REQUEST
-  FETCH_EXPERIMENT_REQUEST,
-  EDIT_EXPERIMENT_NAME_REQUEST,
-  DEPLOY_EXPERIMENT_REQUEST,
-  CREATE_EXPERIMENT_REQUEST,
-  DELETE_EXPERIMENT_REQUEST,
-  FETCH_EXPERIMENT_DEPLOY_STATUS_REQUEST,
-
-  // SUCCESS
-  FETCH_EXPERIMENT_SUCCESS,
-  EDIT_EXPERIMENT_NAME_SUCCESS,
-  DEPLOY_EXPERIMENT_SUCCESS,
-  CREATE_EXPERIMENT_SUCCESS,
-  DELETE_EXPERIMENT_SUCCESS,
-  FETCH_EXPERIMENT_DEPLOY_STATUS_SUCCESS,
-
-  // FAIL
-  FETCH_EXPERIMENT_FAIL,
-  DEPLOY_EXPERIMENT_FAIL,
-  CREATE_EXPERIMENT_FAIL,
-  DELETE_EXPERIMENT_FAIL,
-  FETCH_EXPERIMENT_DEPLOY_STATUS_FAIL,
-
-  // COMMON
-  TRAINING_EXPERIMENT_SUCCEEDED,
-  TRAINING_EXPERIMENT_NOT_SUCCEEDED,
-
-  // DUPLICATE
-  DUPLICATE_EXPERIMENT_FAIL,
-};
