+++ src/store/experiments/actionTypes.js
@@ -2,38 +2,79 @@
 // REQUEST
 // fetch experiments request
 const FETCH_EXPERIMENTS_REQUEST = 'FETCH_EXPERIMENTS_REQUEST';
+// create experiment request
+const CREATE_EXPERIMENT_REQUEST = 'CREATE_EXPERIMENT_REQUEST';
+// update experiment request
+const UPDATE_EXPERIMENT_REQUEST = 'UPDATE_EXPERIMENT_REQUEST';
+// delete experiment request
+const DELETE_EXPERIMENT_REQUEST = 'DELETE_EXPERIMENT_REQUEST';
 // organize experiments request
 const ORGANIZE_EXPERIMENTS_REQUEST = 'ORGANIZE_EXPERIMENTS_REQUEST';
 
 // SUCCESS
 // fetch experiments success
 const FETCH_EXPERIMENTS_SUCCESS = 'FETCH_EXPERIMENTS_SUCCESS';
+// create experiment success
+const CREATE_EXPERIMENT_SUCCESS = 'CREATE_EXPERIMENT_SUCCESS';
+// update experiment success
+const UPDATE_EXPERIMENT_SUCCESS = 'UPDATE_EXPERIMENT_NAME_SUCCESS';
+// delete experiment success
+const DELETE_EXPERIMENT_SUCCESS = 'DELETE_EXPERIMENT_SUCCESS';
 // organize experiments success
 const ORGANIZE_EXPERIMENTS_SUCCESS = 'ORGANIZE_EXPERIMENTS_SUCCESS';
 
 // FAIL
 // fetch experiments fail
 const FETCH_EXPERIMENTS_FAIL = 'FETCH_EXPERIMENTS_FAIL';
+// create experiment fail
+const CREATE_EXPERIMENT_FAIL = 'CREATE_EXPERIMENT_FAIL';
+// update experiment fail
+const UPDATE_EXPERIMENT_FAIL = 'UPDATE_EXPERIMENT_NAME_FAIL';
+// delete experiment fail
+const DELETE_EXPERIMENT_FAIL = 'DELETE_EXPERIMENT_FAIL';
 // organize experiments fail
 const ORGANIZE_EXPERIMENTS_FAIL = 'ORGANIZE_EXPERIMENTS_FAIL';
 
 // UTILS
 const CLEAR_ALL_EXPERIMENTS = 'CLEAR_ALL_EXPERIMENTS';
 
+// DUPLICATE
+const DUPLICATE_EXPERIMENT_FAIL = 'DUPLICATE_EXPERIMENT_FAIL';
+
+// EXPERIMENT STATUS
+const EXPERIMENT_SUCCEDED = 'EXPERIMENT_SUCCEDED';
+const EXPERIMENT_NOT_SUCCEDED = 'EXPERIMENT_NOT_SUCCEDED';
+
 // EXPORT
 export default {
   // REQUEST
   FETCH_EXPERIMENTS_REQUEST,
+  CREATE_EXPERIMENT_REQUEST,
+  UPDATE_EXPERIMENT_REQUEST,
+  DELETE_EXPERIMENT_REQUEST,
   ORGANIZE_EXPERIMENTS_REQUEST,
 
   // SUCCESS
   FETCH_EXPERIMENTS_SUCCESS,
+  CREATE_EXPERIMENT_SUCCESS,
+  UPDATE_EXPERIMENT_SUCCESS,
+  DELETE_EXPERIMENT_SUCCESS,
   ORGANIZE_EXPERIMENTS_SUCCESS,
 
   // FAIL
   FETCH_EXPERIMENTS_FAIL,
+  CREATE_EXPERIMENT_FAIL,
+  UPDATE_EXPERIMENT_FAIL,
+  DELETE_EXPERIMENT_FAIL,
   ORGANIZE_EXPERIMENTS_FAIL,
 
   //CLEAR
   CLEAR_ALL_EXPERIMENTS,
+ 
+  // DUPLICATE
+  DUPLICATE_EXPERIMENT_FAIL,
+
+  // EXPERIMENT STATUS
+  EXPERIMENT_SUCCEDED,
+  EXPERIMENT_NOT_SUCCEDED
 };
