+++ src/store/experiments/index.js
@@ -1,0 +1,10 @@
+import Actions from './actions';
+import ActionTypes from './actionTypes';
+import Reducer from './experimentsReducer';
+
+export default Reducer;
+
+import experimentReducer from './experiment';
+
+export { experimentReducer, Actions, ActionTypes}
+
